spline #1111 Move TxInfo and TxParams types to the model.ts
